requirements.txt
torch
torchvision
scikit-learn
matplotlib
seaborn
pandas
tqdm
pyyaml
numpy
